switched game over div into its own component and added to App.js rather than Header.js
