+++ client/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\index.js":"1","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\App.js":"2","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\Header.js":"3","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\Battleground.js":"4","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\Profile.js":"5","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\Warrior.js":"6","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\utils\\mutations.js":"7","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\utils\\auth.js":"8"},{"size":219,"mtime":1628594069683,"results":"9","hashOfConfig":"10"},{"size":1449,"mtime":1628251226717,"results":"11","hashOfConfig":"10"},{"size":1113,"mtime":1628759034886,"results":"12","hashOfConfig":"10"},{"size":189,"mtime":1628249079201,"results":"13","hashOfConfig":"10"},{"size":3210,"mtime":1628582553692,"results":"14","hashOfConfig":"10"},{"size":238,"mtime":1628230560930,"results":"15","hashOfConfig":"10"},{"size":456,"mtime":1628181089022,"results":"16","hashOfConfig":"10"},{"size":710,"mtime":1628181089022,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"1bmblsf",{"filePath":"21","messages":"22","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"29","messages":"30","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"33","messages":"34","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\index.js",[],["35","36"],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\App.js",[],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\Header.js",[],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\Battleground.js",[],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\Profile.js",[],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\Warrior.js",[],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\utils\\mutations.js",[],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\utils\\auth.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\index.js":"1","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\App.js":"2","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\Header.js":"3","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\Profile.js":"4","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\utils\\mutations.js":"5","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\utils\\auth.js":"6","C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\GameOver.js":"7"},{"size":219,"mtime":1628594069683,"results":"8","hashOfConfig":"9"},{"size":1516,"mtime":1628759636595,"results":"10","hashOfConfig":"9"},{"size":603,"mtime":1628759542883,"results":"11","hashOfConfig":"9"},{"size":3210,"mtime":1628582553692,"results":"12","hashOfConfig":"9"},{"size":456,"mtime":1628181089022,"results":"13","hashOfConfig":"9"},{"size":710,"mtime":1628181089022,"results":"14","hashOfConfig":"9"},{"size":560,"mtime":1628759650127,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"1bmblsf",{"filePath":"19","messages":"20","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"25","messages":"26","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"27","messages":"28","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"29","messages":"30","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\index.js",[],["31","32"],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\App.js",[],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\Header.js",[],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\Profile.js",[],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\utils\\mutations.js",[],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\utils\\auth.js",[],"C:\\Users\\Rob\\code\\Project-3\\immortal-kombat\\client\\src\\components\\GameOver.js",[],{"ruleId":"33","replacedBy":"34"},{"ruleId":"35","replacedBy":"36"},"no-native-reassign",["37"],"no-negated-in-lhs",["38"],"no-global-assign","no-unsafe-negation"]
