--- client/src/components/Battleground.js
@@ -1,11 +1,0 @@
-import React from "react";
-import Profile from "./Profile";
-
-export default function Battleground() {
-  return (
-    <>
-      <Profile />
-      <div id="battleground"></div>
-    </>
-  );
-}
