+++ client/src/components/Header.js
@@ -1,9 +1,4 @@
 function Header({ currentPage, handlePageChange }) {
-  function resetGame() {
-    window.go("game", 0);
-    let highscoreBox = document.querySelector(".scoreboard-container");
-    highscoreBox.setAttribute("class", "score-hidden");
-  }
   return (
     <>
       <header
@@ -22,18 +17,6 @@ function Header({ currentPage, handlePageChange }) {
           </button>
         </a>
       </header>
-
-      <div className="scoreboard-container">
-        <h2 className="game-over">GAME OVER</h2>
-
-        <h5>Score:</h5>
-        <h3 className="score">0</h3>
-        <button className="button">Submit Highscore</button>
-
-        <button className="button" onClick={resetGame}>
-          Play Again
-        </button>
-      </div>
     </>
   );
 }
