--- client/src/components/Warrior.js
@@ -1,10 +1,0 @@
-import React from "react";
-import warriorSprite from "../img/sprites/warrior.png";
-
-export default function Warrior() {
-  return (
-    <>
-      <img className="warriorImg" src={warriorSprite} alt="warrior"></img>
-    </>
-  );
-}
